% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{ensureExampleFiles,character,RappCoreExamplesS3,RappCoreExamplesS3-method}
\alias{ensureExampleFiles,character,RappCoreExamplesS3,RappCoreExamplesS3-method}
\title{Ensure Example Files (\code{character-RappCoreExamplesS3-RappCoreExamplesS3})}
\usage{

  \S4method{ensureExampleFiles}{character,RappCoreExamplesS3,RappCoreExamplesS3}(pkg = ".",
  ..., .ctx, .ns)
}
\arguments{
\item{pkg}{\code{\link{character}}}

\item{.ctx}{\code{\link{RappCoreExamplesS3}}.}

\item{.ns}{\code{\link{RappCoreExamplesS3}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method:
  	\code{\link[rapp.core.examples]{ensureExampleFiles-character-RappCoreExamplesS3-RappCoreExamplesS3-method}}.
}
\description{
See generic \link[rapp.core.examples]{getBinaryExampleStatus}
}
\details{
Note that you should apply this method \emph{after} a fresh initialization
of your \package{rapp} project as this will ensure that dummy example files
are created for all of your functions/methods.
}
\examples{
\dontrun{

ensureExampleFiles()

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp/rapp.core.examples}
}
\seealso{
\code{
		\link[rapp.core.examples]{ensureExampleFiles}
}
}

