% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ensureExampleFiles}
\alias{ensureExampleFiles}
\title{Ensure Example Files}
\usage{
ensureExampleFiles(pkg = ".", ..., .ctx, .ns)
}
\arguments{
\item{pkg}{\strong{Signature argument}.
Package name or path to R package project.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}

\item{.ctx}{\strong{Signature argument}. Application context.}

\item{.ns}{\strong{Signature argument}. Application namespace.}
}
\description{
Ensures that example files specified via \code{@example} exist.
}
\section{General Remarks}{

\subsection{Arguments \code{.ctx} and \code{.ns}}{
 Take into account possible non-missing values for \code{.ctx} and \code{.ns}
 depending which application context and/or namespace you want to use the
 package in when you call methods of this package.
 }
}
\examples{
\dontrun{

ensureExampleFiles()

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp/rapp.core.examples}
}
\seealso{
\code{
   \link[rapp2]{ensureExampleFiles-character-RappCoreExamplesS3-RappCoreExamplesS3-method}
}
}

