% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getBinaryExampleStatus,character,missing,missing-method}
\alias{getBinaryExampleStatus,character,missing,missing-method}
\title{Get Binary Example Status (\code{character-missing-missing)}}
\usage{
\S4method{getBinaryExampleStatus}{character,missing,missing}(path,
  pattern = "## TODO: add example", ..., .ctx, .ns)
}
\arguments{
\item{path}{\code{\link{character}}. Dimension: 1.
Directory path (directory that contains
  the example cript files).}

\item{.ctx}{\code{\link{missing}}.}

\item{.ns}{\code{\link{missing}}.}

\item{pattern}{\code{\link{character}}. Dimension: 1.
Pattern to recoginize empty example files.
Refactoring argument, do not change.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method:
		\code{\link[rapp.core.examples]{getBinaryExampleStatus-character-RappCoreExamplesS3-RappCoreExamplesS3-method}}.
}
\description{
Get Binary Example Status (\code{character-missing-missing)}
}
\section{General Remarks}{

\subsection{Arguments \code{.ctx} and \code{.ns}}{
 Take into account possible non-missing values for \code{.ctx} and \code{.ns}
 depending which application context and/or namespace you want to use the
 package in when you call methods of this package.
 }
}
\examples{
\dontrun{

## Assuming you are calling from within a package project //
getBinaryExampleStatus(path = "inst/examples")

try(getBinaryExampleStatus(path = "non-existing-path"))

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp/rapp.core.examples}
}
\seealso{
\code{
		\link[rapp.core.examples]{getBinaryExampleStatus},
		\link[rapp.core.examples]{getBinaryExampleStatus-character-RappCoreExamplesS3-RappCoreExamplesS3-method}
}
}

