% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getBinaryExampleStatus,character,RappCoreExamplesS3,RappCoreExamplesS3-method}
\alias{getBinaryExampleStatus,character,RappCoreExamplesS3,RappCoreExamplesS3-method}
\title{Get Binary Example Status (\code{character-RappCoreExamplesS3-RappCoreExamplesS3)}}
\usage{

  \S4method{getBinaryExampleStatus}{character,RappCoreExamplesS3,RappCoreExamplesS3}(path,
  pattern = "## TODO: add example", ..., .ctx, .ns)
}
\arguments{
\item{path}{\code{\link{character}}. Dimension: 1.
Directory path (directory that contains
    the example cript files).}

\item{.ctx}{\code{\link{RappCoreExamplesS3}}.}

\item{.ns}{\code{\link{RappCoreExamplesS3}}.}

\item{pattern}{\code{\link{character}}. Dimension: 1.
Pattern to recoginize empty example files.
Refactoring argument, do not change.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{logical}. Logical vector whose length corresponds to the
   number of files identified in \code{path}. For each example file,
   \code{TRUE} means actual example code exists, \code{FALSE} means it does not.
}
\description{
Get Binary Example Status (\code{character-RappCoreExamplesS3-RappCoreExamplesS3)}
}
\section{General Remarks}{

\subsection{Arguments \code{.ctx} and \code{.ns}}{
 Take into account possible non-missing values for \code{.ctx} and \code{.ns}
 depending which application context and/or namespace you want to use the
 package in when you call methods of this package.
 }
}
\examples{
\dontrun{

## Assuming you are calling from within a package project //
getBinaryExampleStatus(path = "inst/examples")

try(getBinaryExampleStatus(path = "non-existing-path"))

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp/rapp.core.examples}
}
\seealso{
\code{
		\link[rapp.core.examples]{getBinaryExampleStatus},
		\link[rapp.core.examples]{getBinaryExampleStatus-character-missing-missing-method}
}
}

