% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getBinaryExampleStatus}
\alias{getBinaryExampleStatus}
\title{Get Binary Example Status}
\usage{
getBinaryExampleStatus(path, pattern = "## TODO: add example", ..., .ctx, .ns)
}
\arguments{
\item{path}{\strong{Signature argument}.
  Object containing directory path information
(directory that contains the example script files).}

\item{pattern}{\code{\link{character}}. Dimension: 1.
Pattern to recoginize empty example files.
Refactoring argument, do not change.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}

\item{.ctx}{\strong{Signature argument}. Application context.}

\item{.ns}{\strong{Signature argument}. Application namespace.}
}
\description{
Checks for each function/method if an non-empty example has been been
specified or not.
}
\section{General Remarks}{

\subsection{Arguments \code{.ctx} and \code{.ns}}{
 Take into account possible non-missing values for \code{.ctx} and \code{.ns}
 depending which application context and/or namespace you want to use the
 package in when you call methods of this package.
 }
}
\examples{
\dontrun{

## Assuming you are calling from within a package project //
getBinaryExampleStatus(path = "inst/examples")

try(getBinaryExampleStatus(path = "non-existing-path"))

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp/rapp.core.examples}
}
\seealso{
\code{
 \link[rapp.core.examples]{getBinaryExampleStatus-character-missing-missing-method},
 \link[rapp.core.examples]{getBinaryExampleStatus-character-RappCoreExamplesS3-RappCoreExamplesS3-method}
}
}

