% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getExistingExamples,character,RappCoreExamplesS3,RappCoreExamplesS3-method}
\alias{getExistingExamples,character,RappCoreExamplesS3,RappCoreExamplesS3-method}
\title{Get Existing Example (\code{character-RappCoreExamplesS3-RappCoreExamplesS3})}
\usage{

  \S4method{getExistingExamples}{character,RappCoreExamplesS3,RappCoreExamplesS3}(path,
  pattern = "## TODO: add example", include_fileext = FALSE,
  return_path = FALSE, ..., .ctx, .ns)
}
\arguments{
\item{path}{\code{\link{character}}. Dimension: 1.
Directory path (directory that contains
    the example script files).}

\item{include_fileext}{\code{\link{logical}}. Dimension: 1.
\code{TRUE} means return value includes file extensions,
\code{FALSE} means file extensions are removed.
Only considered if \code{return_path = FALSE}. Otherwise it is automatically
set to \code{TRUE}.}

\item{return_path}{\code{\link{logical}}. Dimension: 1.
\code{TRUE} means return value corresponds to paths of example files,
\code{FALSE} (default) means return value corresponds to
function/method names.}

\item{.ctx}{\code{\link{missing}}.}

\item{.ns}{\code{\link{missing}}.}

\item{pattern}{See \code{pattern} of
\code{\link[rapp.core.examples]{getBinaryExampleStatus}}}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
TODO
}
\description{
See generic \link[rapp.core.examples]{getBinaryExampleStatus}
}
\details{
Note that you should apply this method \emph{after} a fresh initialization
of your \package{rapp} project as this will ensure that dummy example files
are created for all of your functions/methods.
}
\examples{
\dontrun{

## Assuming you are calling from within a package project //
getExistingExamples(path = "inst/examples")

try(getExistingExamples(path = "non-existing-path"))

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp/rapp.core.examples}
}
\seealso{
\code{
		\link[rapp.core.examples]{getExistingExamples},
		\link[rapp.core.examples]{getExistingExamples-character-missing-missing-method}
}
}

