% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getExistingExamples}
\alias{getExistingExamples}
\title{Get Existing Example}
\usage{
getExistingExamples(path, pattern = "## TODO: add example",
  include_fileext = FALSE, return_path = FALSE, ..., .ctx, .ns)
}
\arguments{
\item{path}{\strong{Signature argument}.
Object containing directory path information (directory that contains
the examples' script files).}

\item{pattern}{See \code{pattern} of
\code{\link[rapp.core.examples]{getBinaryExampleStatus}}}

\item{include_fileext}{\code{\link{logical}}. Dimension: 1.
\code{TRUE} means return value includes file extensions,
  \code{FALSE} means file extensions are removed.
  Only considered if \code{return_path = FALSE}. Otherwise it is automatically
  set to \code{TRUE}.}

\item{return_path}{\code{\link{logical}}. Dimension: 1.
\code{TRUE} means return value corresponds to paths of example files,
\code{FALSE} (default) means return value corresponds to
function/method names.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}

\item{.ctx}{\strong{Signature argument}. Application context.}

\item{.ns}{\strong{Signature argument}. Application namespace.}
}
\description{
Gets names of functions/methods for which examples already exist.
}
\section{General Remarks}{

\subsection{Arguments \code{.ctx} and \code{.ns}}{
 Take into account possible non-missing values for \code{.ctx} and \code{.ns}
 depending which application context and/or namespace you want to use the
 package in when you call methods of this package.
 }
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp/rapp.core.examples}
}
\seealso{
\code{
   \link[rapp.core.examples]{getExistingExamples-character-missing-missing-method},
   \link[rapp.core.examples]{getExistingExamples-character-RappCoreExamplesS3-RappCoreExamplesS3-method},
   \link[rapp.core.examples]{getExampleBinaryStatus},
		\link[rapp.core.examples]{getMissingExamples}
}
}

