% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getMissingExamples,character,RappCoreExamplesS3,RappCoreExamplesS3-method}
\alias{getMissingExamples,character,RappCoreExamplesS3,RappCoreExamplesS3-method}
\title{Get Missing Examples (\code{character-RappCoreExamplesS3-RappCoreExamplesS3})}
\usage{

  \S4method{getMissingExamples}{character,RappCoreExamplesS3,RappCoreExamplesS3}(path,
  pattern = "## TODO: add example", include_fileext = FALSE,
  return_path = FALSE, ..., .ctx, .ns)
}
\arguments{
\item{path}{\code{\link{character}}. Dimension: 1.
Directory path (directory that contains
  the examples' script files).}

\item{.ctx}{\code{\link{RappCoreExamplesS3}}.}

\item{.ns}{\code{\link{RappCoreExamplesS3}}.}

\item{pattern}{See \code{\link[rapp.core.examples]{getBinaryExampleStatus}}.}

\item{include_fileext}{See \code{\link[rapp.core.examples]{getExistingExamples}}.}

\item{return_path}{See \code{\link[rapp.core.examples]{getExistingExamples}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{character}}. Names of functions/methods for which
		there exist no examples yet.
}
\description{
See generic \link[rapp.core.examples]{getBinaryExampleStatus}
}
\details{
Note that you should apply this method \emph{after} a fresh initialization
of your \package{rapp} project as this will ensure that dummy example files
are created for all of your functions/methods.
}
\examples{
\dontrun{

## Assuming you are calling from within a package project //
getMissingExamples(path = "inst/examples")

try(getMissingExamples(path = "non-existing-path"))

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp/rapp.core.examples}
}
\seealso{
\code{
		\link[rapp.core.examples]{getBinaryExampleStatus},
		\link[rapp.core.examples]{getMissingExamples-character-RappCoreExamplesS3-RappCoreExamplesS3-method}
}
}

