% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getMissingExamples}
\alias{getMissingExamples}
\title{Get Missing Examples}
\usage{
getMissingExamples(path, pattern = "## TODO: add example",
  include_fileext = FALSE, return_path = FALSE, ..., .ctx, .ns)
}
\arguments{
\item{path}{\strong{Signature argument}.
Object containing directory path information (directory that contains
the examples' script files).}

\item{pattern}{See \code{\link[rapp.core.examples]{getBinaryExampleStatus}}.}

\item{include_fileext}{See \code{\link[rapp.core.examples]{getExistingExamples}}.}

\item{return_path}{See \code{\link[rapp.core.examples]{getExistingExamples}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}

\item{.ctx}{\strong{Signature argument}. Application context.}

\item{.ns}{\strong{Signature argument}. Application namespace.}
}
\description{
Gets names of functions/methods for which examples are still missing.
}
\section{General Remarks}{

\subsection{Arguments \code{.ctx} and \code{.ns}}{
 Take into account possible non-missing values for \code{.ctx} and \code{.ns}
 depending which application context and/or namespace you want to use the
 package in when you call methods of this package.
 }
}
\examples{
\dontrun{

## Assuming you are calling from within a package project //
getMissingExamples(path = "inst/examples")

try(getMissingExamples(path = "non-existing-path"))

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp/rapp.core.examples}
}
\seealso{
\code{
   \link[rapp.core.examples]{getMissingExamples-character-missing-missing-method},
   \link[rapp.core.examples]{getMissingExamples-character-RappCoreExamplesS3-RappCoreExamplesS3-method},
		\link[rapp.core.examples]{getExampleBinaryStatus},
		\link[rapp.core.examples]{getExistingExamples}
}
}

