% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{isDontRunExample,character,missing,missing-method}
\alias{isDontRunExample,character,missing,missing-method}
\title{Check for Don't-Run Examples (\code{character-missing-missing})}
\usage{
\S4method{isDontRunExample}{character,missing,missing}(path,
  pattern = "## TODO: add example", ..., .ctx, .ns)
}
\arguments{
\item{path}{\code{\link{character}}. Dimension: 1.
Directory path (directory that contains
  the examples' script files).}

\item{.ctx}{\code{\link{missing}}.}

\item{.ns}{\code{\link{missing}}.}

\item{pattern}{See \code{pattern} of
\code{\link[rapp.core.examples]{getBinaryExampleStatus}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method:
		\code{\link[rapp.core.examples]{isDontRunExample-character-RappCoreExamplesS3-RappCoreExamplesS3-method}}.
}
\description{
See generic \link[rapp.core.examples]{getBinaryExampleStatus}
}
\examples{
\dontrun{

## Assuming you are calling from within a package project //
isDontRunExample(path = "inst/examples")

try(isDontRunExample(path = "non-existing-path"))

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp/rapp.core.examples}
}
\seealso{
\code{
		\link[rapp.core.examples]{isDontRunExample},
		\link[rapp.core.examples]{isDontRunExample-character-RappCoreExamplesS3-RappCoreExamplesS3-method}
}
}

