% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{isDontRunExample,character,RappCoreExamplesS3,RappCoreExamplesS3-method}
\alias{isDontRunExample,character,RappCoreExamplesS3,RappCoreExamplesS3-method}
\title{Check for Don't-Run Examples (\code{character-RappCoreExamplesS3-RappCoreExamplesS3})}
\usage{

  \S4method{isDontRunExample}{character,RappCoreExamplesS3,RappCoreExamplesS3}(path,
  pattern = "## TODO: add example", ..., .ctx, .ns)
}
\arguments{
\item{path}{\code{\link{character}}. Dimension: 1.
Directory path (directory that contains
  the examples' script files).}

\item{.ctx}{\code{\link{RappCoreExamplesS3}}.}

\item{.ns}{\code{\link{RappCoreExamplesS3}}.}

\item{pattern}{See \code{pattern} of
\code{\link[rapp.core.examples]{getBinaryExampleStatus}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{logical}}. \code{TRUE} means example is a
		"dont-run" example, \code{FALSE} means it is not.
}
\description{
See generic \link[rapp.core.examples]{getBinaryExampleStatus}
}
\details{
Note that you should apply this method \emph{after} a fresh initialization
of your \package{rapp} project as this will ensure that dummy example files
are created for all of your functions/methods.
}
\examples{
\dontrun{

## Assuming you are calling from within a package project //
isDontRunExample(path = "inst/examples")

try(isDontRunExample(path = "non-existing-path"))

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp/rapp.core.examples}
}
\seealso{
\code{
		\link[rapp.core.examples]{isDontRunExample},
		\link[rapp.core.examples]{isDontRunExample-character-missing-missing-method}
}
}

