% Generated by roxygen2 (4.0.1): do not edit by hand
\name{isDontRunExample}
\alias{isDontRunExample}
\title{Check for Don't-Run Examples}
\usage{
isDontRunExample(path, pattern = "## TODO: add example", ..., .ctx, .ns)
}
\arguments{
\item{path}{\strong{Signature argument}.
Object containing directory path information (directory that contains
the examples' script files).}

\item{pattern}{See \code{pattern} of
\code{\link[rapp.core.examples]{getBinaryExampleStatus}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}

\item{.ctx}{\strong{Signature argument}. Application context.}

\item{.ns}{\strong{Signature argument}. Application namespace.}
}
\description{
Checks for each example files if the respective example is a "don't-run"
example or not.
}
\section{General Remarks}{

\subsection{Arguments \code{.ctx} and \code{.ns}}{
 Take into account possible non-missing values for \code{.ctx} and \code{.ns}
 depending which application context and/or namespace you want to use the
 package in when you call methods of this package.
 }
}
\examples{
\dontrun{

## Assuming you are calling from within a package project //
isDontRunExample(path = "inst/examples")

try(isDontRunExample(path = "non-existing-path"))

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp/rapp.core.examples}
}
\seealso{
\code{
   \link[rapp.core.examples]{isDontRunExample-character-missing-missing-method},
   \link[rapp.core.examples]{isDontRunExample-character-RappCoreExamplesS3-RappCoreExamplesS3-method},
}
}

