% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_iteration.R
\name{root_to_tip}
\alias{root_to_tip}
\title{Root-to-tip calculates the distance between the root and each node}
\usage{
root_to_tip(x)
}
\arguments{
\item{x}{Tree of class 'phylo'}
}
\value{
A vector of root distances (numeric)
}
\description{
Root-to-tip calculates distance from root of phylogenetic tree object
to each tip. Returns a vector of distances.
}
\examples{
sample_tree <- ape::rtree(5)
root_to_tip(sample_tree)
}
