% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{runExamples,character,RappCoreExamplesS3,RappCoreExamplesS3-method}
\alias{runExamples,character,RappCoreExamplesS3,RappCoreExamplesS3-method}
\title{Run Examples (\code{character-RappCoreExamplesS3-RappCoreExamplesS3})}
\usage{
\S4method{runExamples}{character,RappCoreExamplesS3,RappCoreExamplesS3}(path,
  dontrun = FALSE, ..., .ctx, .ns)
}
\arguments{
\item{path}{\code{character}. Dimension: 1.
Directory path.}

\item{.ctx}{\code{\link{RappCoreExamplesS3}}.}

\item{.ns}{\code{\link{RappCoreExamplesS3}}.}

\item{dontrun}{\code{\link{logical}}. Dimension: 1.
\code{TRUE} means dont-run examples are included,
\code{FALSE} (default) means only non-dont-run examples are considered.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method:
		\code{\link[rapp.core.examples]{runExamples-character-RappCoreExamplesS3-RappCoreExamplesS3}}
}
\description{
Run Examples (\code{character-RappCoreExamplesS3-RappCoreExamplesS3})
}
\section{General Remarks}{

\subsection{Arguments \code{.ctx} and \code{.ns}}{
 Take into account possible non-missing values for \code{.ctx} and \code{.ns}
 depending which application context and/or namespace you want to use the
 package in when you call methods of this package.
 }
}
\examples{
\dontrun{

# TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp/rapp.core.examples}
}

