% Generated by roxygen2 (4.0.1): do not edit by hand
\name{runExamples}
\alias{runExamples}
\title{Run Examples}
\usage{
runExamples(path, dontrun = FALSE, ..., .ctx, .ns)
}
\arguments{
\item{path}{\strong{Signature argument}.
Object containing directory path information
(directory that contains the example script files).}

\item{dontrun}{\code{\link{logical}}. Dimension: 1.
\code{TRUE} means dont-run examples are included,
\code{FALSE} (default) means only non-dont-run examples are considered.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}

\item{.ctx}{\strong{Signature argument}. Application context.}

\item{.ns}{\strong{Signature argument}. Application namespace.}
}
\description{
Run examples in \code{inst/examples}.
}
\details{
Note that you should apply this method \emph{after} a fresh initialization
of your \package{rapp} project as this will ensure that dummy example files
are created for all of your functions/methods.
}
\section{General Remarks}{

\subsection{Arguments \code{.ctx} and \code{.ns}}{
 Take into account possible non-missing values for \code{.ctx} and \code{.ns}
 depending which application context and/or namespace you want to use the
 package in when you call methods of this package.
 }
}
\examples{
\dontrun{

# TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp/rapp.core.examples}
}

